Move shared types to osu.Game project
