+++ osu.Server.Spectator/Database/IDatabaseAccess.cs
@@ -4,9 +4,9 @@
 using System;
 using System.IdentityModel.Tokens.Jwt;
 using System.Threading.Tasks;
+using osu.Game.Online.Metadata;
 using osu.Game.Online.Multiplayer;
 using osu.Server.Spectator.Database.Models;
-using osu.Server.Spectator.Hubs;
 
 namespace osu.Server.Spectator.Database
 {
